+++ cosmwasm/packages/registry/src/error.rs
@@ -15,6 +15,9 @@ pub enum RegistryError {
     #[error("{0}")]
     ValueSerialization(ValueSerError),
 
+    #[error("{0}")]
+    Bech32ErrorRaw(#[from] bech32::Error),
+
     // Validation errors
     #[error("Invalid channel id: {0}")]
     InvalidChannelId(String),
@@ -29,6 +32,9 @@ pub enum RegistryError {
 
     #[error("serialization error: {error}")]
     SerialiaztionError { error: String },
+
+    #[error("registry improperly configured")]
+    ImproperlyConfigured {},
 
     #[error("denom {denom:?} is not an IBC denom")]
     InvalidIBCDenom { denom: String },
@@ -69,20 +75,14 @@ pub enum RegistryError {
     #[error("no authorized address found for source chain: {source_chain:?}")]
     ChainAuthorizedAddressDoesNotExist { source_chain: String },
 
-    #[error("chain channel link does not exist: {source_chain:?} -> {destination_chain:?}")]
+    #[error("channel between chains not registered: {source_chain:?} -> {destination_chain:?}")]
     ChainChannelLinkDoesNotExist {
         source_chain: String,
         destination_chain: String,
     },
 
-    #[error("channel chain link does not exist: {channel_id:?} on {source_chain:?} -> chain")]
-    ChannelChainLinkDoesNotExist {
-        channel_id: String,
-        source_chain: String,
-    },
-
-    #[error("channel chain link does not exist: {channel_id:?} on {source_chain:?} -> chain")]
-    ChannelToChainChainLinkDoesNotExist {
+    #[error("channel to chain link not registered: {channel_id:?} on {source_chain:?}")]
+    ChannelDoesNotExistOnChain {
         channel_id: String,
         source_chain: String,
     },
@@ -92,6 +92,9 @@ pub enum RegistryError {
 
     #[error("bech32 prefix does not exist for chain: {chain}")]
     Bech32PrefixDoesNotExist { chain: String },
+
+    #[error("Chain {chain} does not support forwarding")]
+    ForwardingUnsopported { chain: String },
 }
 
 impl From<RegistryError> for StdError {