+++ cosmwasm/packages/registry/src/registry.rs
@@ -19,6 +19,44 @@ pub fn hash_denom_trace(unwrapped: &str) -> String {
     format!("ibc/{}", hash.to_uppercase())
 }
 
+// When a contract is called using ibc callbacks, the addres is a combination of
+// the channel and the original sender. This function lets us compute that.
+pub fn derive_wasmhooks_sender(
+    channel: &str,
+    original_sender: &str,
+    bech32_prefix: &str,
+) -> Result<String, RegistryError> {
+    let sender = format!("{channel}/{original_sender}");
+
+    let mut hasher0 = Sha256::new();
+    hasher0.update("ibc-wasm-hook-intermediary".as_bytes());
+    let th = hasher0.finalize();
+
+    let mut hasher = Sha256::new();
+    hasher.reset();
+    hasher.update(th.as_slice());
+    hasher.update(sender.as_bytes());
+
+    let result = hasher.finalize();
+
+    // The bech32 crate requires a Vec<u5> as input, so we need to convert the bytes.
+    let result_u5 = bech32::convert_bits(result.as_slice(), 8, 5, true)?;
+    // result_u5 contains the bytes as a u5 but in an u8 type, so we need to explicitly
+    // do the type conversion
+    let result_u5: Vec<bech32::u5> = result_u5
+        .iter()
+        .filter_map(|&x| bech32::u5::try_from_u8(x).ok())
+        .collect();
+
+    bech32::encode(bech32_prefix, result_u5, bech32::Variant::Bech32).map_err(|e| {
+        RegistryError::Bech32Error {
+            action: "encoding".into(),
+            addr: original_sender.into(),
+            source: e,
+        }
+    })
+}
+
 // IBC transfer port
 const TRANSFER_PORT: &str = "transfer";
 // IBC timeout
@@ -101,6 +139,9 @@ pub struct MultiHopDenom {
     pub via: Option<ChannelId>, // This is optional because native tokens have no channel
 }
 
+// The name of the chain on which the contract using this lib is instantiated
+pub const CONTRACT_CHAIN: &str = "osmosis";
+
 pub struct Registry<'a> {
     pub deps: Deps<'a>,
     pub registry_contract: String,
@@ -115,7 +156,6 @@ impl<'a> Registry<'a> {
         })
     }
 
-    #[allow(dead_code)]
     pub fn default(deps: Deps<'a>) -> Self {
         Self {
             deps,
@@ -163,7 +203,7 @@ impl<'a> Registry<'a> {
                     via_channel: via_channel.to_string(),
                 },
             )
-            .map_err(|_e| RegistryError::ChannelToChainChainLinkDoesNotExist {
+            .map_err(|_e| RegistryError::ChannelDoesNotExistOnChain {
                 channel_id: via_channel.to_string(),
                 source_chain: on_chain.to_string(),
             })
@@ -186,6 +226,20 @@ impl<'a> Registry<'a> {
                 source_chain: on_chain.to_string(),
                 destination_chain: for_chain.to_string(),
             })
+    }
+
+    /// Returns a boolean specifying if a chain supports forwarding
+    /// Example: supports_forwarding("gaia") -> true
+    pub fn supports_forwarding(&self, chain: &str) -> Result<bool, RegistryError> {
+        self.deps
+            .querier
+            .query_wasm_smart(
+                &self.registry_contract,
+                &QueryMsg::HasPacketForwarding {
+                    chain: chain.to_string(),
+                },
+            )
+            .map_err(|_e| RegistryError::ImproperlyConfigured {})
     }
 
     /// Re-encodes the bech32 address for the receiving chain
@@ -262,7 +316,7 @@ impl<'a> Registry<'a> {
     pub fn unwrap_denom_path(&self, denom: &str) -> Result<Vec<MultiHopDenom>, RegistryError> {
         self.debug(format!("Unwrapping denom {denom}"));
 
-        let mut current_chain = "osmosis".to_string(); // The initial chain is always osmosis
+        let mut current_chain = CONTRACT_CHAIN.to_string(); // The initial chain is always the contract chain
 
         // Check that the denom is an IBC denom
         if !denom.starts_with("ibc/") {
@@ -289,9 +343,7 @@ impl<'a> Registry<'a> {
             }),
         }?;
 
-        self.deps
-            .api
-            .debug(&format!("procesing denom trace {path}"));
+        self.debug(format!("procesing denom trace {path}"));
         // Let's iterate over the parts of the denom trace and extract the
         // chain/channels into a more useful structure: MultiHopDenom
         let mut hops: Vec<MultiHopDenom> = vec![];
@@ -331,6 +383,21 @@ impl<'a> Registry<'a> {
         Ok(hops)
     }
 
+    pub fn get_native_chain(&self, denom: &str) -> Result<Chain, RegistryError> {
+        let hops = self.unwrap_denom_path(denom)?;
+        self.debug(format!("hops: {hops:?}"));
+        // verify that the last hop is native
+        let last_hop = hops.last().ok_or(RegistryError::NoDenomTrace {
+            denom: denom.into(),
+        })?;
+        if last_hop.via.is_some() {
+            return Err(RegistryError::InvalidDenomTrace {
+                error: format!("Path {hops:?} is not properly formatted"),
+            });
+        }
+        Ok(hops.last().unwrap().on.clone())
+    }
+
     /// Returns an IBC MsgTransfer that with a packet forward middleware memo
     /// that will send the coin back to its original chain and then to the
     /// receiver in `into_chain`.
@@ -354,6 +421,7 @@ impl<'a> Registry<'a> {
         block_time: Timestamp,
         first_transfer_memo: String,
         receiver_callback: Option<Callback>,
+        skip_forwarding_check: bool,
     ) -> Result<proto::MsgTransfer, RegistryError> {
         // Calculate the path that this coin took to get to the current chain.
         // Each element in the path is an IBC hop.
@@ -470,6 +538,17 @@ impl<'a> Registry<'a> {
                 None => ChannelId(self.get_channel(prev_chain, hop.on.as_ref())?),
             };
 
+            self.debug(format!(
+                "checking that: {:?} supports pfm (into {:?})",
+                hop.on.as_ref(),
+                prev_chain
+            ));
+            if !skip_forwarding_check && !self.supports_forwarding(hop.on.as_ref())? {
+                return Err(RegistryError::ForwardingUnsopported {
+                    chain: hop.on.as_ref().into(),
+                });
+            }
+
             // The next memo wraps the previous one
             next = Some(Box::new(Memo {
                 forward: Some(ForwardingMemo {
@@ -576,5 +655,40 @@ mod test {
             encoded,
             r#"{"forward":{"receiver":"receiver","port":"port","channel":"channel-0","next":{"forward":{"receiver":"receiver2","port":"port2","channel":"channel-1"}}}}"#
         )
+    }
+
+    #[test]
+    fn test_derive_wasmhooks_sender() {
+        let test_cases = vec![
+            (
+                "channel-0",
+                "cosmos1tfejvgp5yzd8ypvn9t0e2uv2kcjf2laa8upya8",
+                "osmo",
+                "osmo1sguz3gtyl2tjsdulwxmtprd68xtd43yyep6g5c554utz642sr8rqcgw0q6",
+            ),
+            (
+                "channel-1",
+                "cosmos1tfejvgp5yzd8ypvn9t0e2uv2kcjf2laa8upya8",
+                "osmo",
+                "osmo1svnare87kluww5hnltv24m4dg72hst0qqwm5xslsvnwd22gftcussaz5l7",
+            ),
+            (
+                "channel-0",
+                "osmo12smx2wdlyttvyzvzg54y2vnqwq2qjateuf7thj",
+                "osmo",
+                "osmo1vz8evs4ek3vnz4f8wy86nw9ayzn67y28vtxzjgxv6achc4pa8gesqldfz0",
+            ),
+            (
+                "channel-0",
+                "osmo12smx2wdlyttvyzvzg54y2vnqwq2qjateuf7thj",
+                "cosmos",
+                "cosmos1vz8evs4ek3vnz4f8wy86nw9ayzn67y28vtxzjgxv6achc4pa8ges4z434f",
+            ),
+        ];
+
+        for tc in test_cases {
+            assert!(derive_wasmhooks_sender(tc.0, tc.1, tc.2).is_ok());
+            assert_eq!(derive_wasmhooks_sender(tc.0, tc.1, tc.2).unwrap(), tc.3);
+        }
     }
 }