+++ x/protorev/keeper/routes_test.go
@@ -44,12 +44,16 @@ func (s *KeeperTestSuite) TestBuildRoutes() {
 			description: "Route exists for swap in Bitcoin and swap out Atom",
 			inputDenom:  "bitcoin",
 			outputDenom: "Atom",
-			poolID:      4,
+			poolID:      55,
 			expectedRoutes: [][]TestRoute{
 				{
 					{PoolId: 25, InputDenom: types.OsmosisDenomination, OutputDenom: "Atom"},
-					{PoolId: 4, InputDenom: "Atom", OutputDenom: "bitcoin"},
+					{PoolId: 55, InputDenom: "Atom", OutputDenom: "bitcoin"},
 					{PoolId: 10, InputDenom: "bitcoin", OutputDenom: types.OsmosisDenomination},
+				},
+				{
+					{PoolId: 55, InputDenom: "Atom", OutputDenom: "bitcoin"},
+					{PoolId: 4, InputDenom: "bitcoin", OutputDenom: "Atom"},
 				},
 			},
 		},
@@ -91,6 +95,22 @@ func (s *KeeperTestSuite) TestBuildRoutes() {
 				},
 			},
 		},
+		{
+			description: "Two Pool Route exists for (osmo, atom)",
+			inputDenom:  "Atom",
+			outputDenom: types.OsmosisDenomination,
+			poolID:      51,
+			expectedRoutes: [][]TestRoute{
+				{
+					{PoolId: 51, InputDenom: types.OsmosisDenomination, OutputDenom: "Atom"},
+					{PoolId: 25, InputDenom: "Atom", OutputDenom: types.OsmosisDenomination},
+				},
+				{
+					{PoolId: 25, InputDenom: "Atom", OutputDenom: types.OsmosisDenomination},
+					{PoolId: 51, InputDenom: types.OsmosisDenomination, OutputDenom: "Atom"},
+				},
+			},
+		},
 	}
 
 	for _, tc := range cases {
@@ -101,6 +121,7 @@ func (s *KeeperTestSuite) TestBuildRoutes() {
 			for routeIndex, route := range routes {
 				for tradeIndex, poolID := range route.Route.PoolIds() {
 					s.Require().Equal(tc.expectedRoutes[routeIndex][tradeIndex].PoolId, poolID)
+					s.Require().Equal(tc.expectedRoutes[routeIndex][tradeIndex].OutputDenom, route.Route[tradeIndex].TokenOutDenom)
 				}
 			}
 		})
@@ -215,6 +236,99 @@ func (s *KeeperTestSuite) TestBuildHighestLiquidityRoute() {
 					s.Require().Equal(trade.PoolId, routeMetaData.Route.PoolIds()[index])
 				}
 			} else {
+				s.Require().Error(err)
+			}
+		})
+	}
+}
+
+// TestBuildTwoPoolRoute tests the BuildTwoPoolRoute function
+func (s *KeeperTestSuite) TestBuildTwoPoolRoute() {
+	cases := []struct {
+		description   string
+		swapDenom     types.BaseDenom
+		tokenIn       string
+		tokenOut      string
+		poolId        uint64
+		expectedRoute []TestRoute
+		hasRoute      bool
+	}{
+		{
+			description: "two pool route can be created with base as token out",
+			swapDenom: types.BaseDenom{
+				Denom:    types.OsmosisDenomination,
+				StepSize: sdk.NewInt(1_000_000),
+			},
+			tokenIn:  "stake",
+			tokenOut: types.OsmosisDenomination,
+			poolId:   54,
+			expectedRoute: []TestRoute{
+				{PoolId: 54, InputDenom: types.OsmosisDenomination, OutputDenom: "stake"},
+				{PoolId: 55, InputDenom: "stake", OutputDenom: types.OsmosisDenomination},
+			},
+			hasRoute: true,
+		},
+		{
+			description: "two pool route can be created with base as token in",
+			swapDenom: types.BaseDenom{
+				Denom:    types.OsmosisDenomination,
+				StepSize: sdk.NewInt(1_000_000),
+			},
+			tokenIn:  types.OsmosisDenomination,
+			tokenOut: "stake",
+			poolId:   54,
+			expectedRoute: []TestRoute{
+				{PoolId: 55, InputDenom: types.OsmosisDenomination, OutputDenom: "stake"},
+				{PoolId: 54, InputDenom: "stake", OutputDenom: types.OsmosisDenomination},
+			},
+			hasRoute: true,
+		},
+		{
+			description: "two pool route where swap is on the highest liquidity pool",
+			swapDenom: types.BaseDenom{
+				Denom:    types.OsmosisDenomination,
+				StepSize: sdk.NewInt(1_000_000),
+			},
+			tokenIn:       "stake",
+			tokenOut:      types.OsmosisDenomination,
+			poolId:        55,
+			expectedRoute: []TestRoute{},
+			hasRoute:      false,
+		},
+		{
+			description: "trade executes on pool not tracked by the module",
+			swapDenom: types.BaseDenom{
+				Denom:    types.OsmosisDenomination,
+				StepSize: sdk.NewInt(1_000_000),
+			},
+			tokenIn:       "stake",
+			tokenOut:      types.OsmosisDenomination,
+			poolId:        1000000,
+			expectedRoute: []TestRoute{},
+			hasRoute:      false,
+		},
+	}
+
+	for _, tc := range cases {
+		s.Run(tc.description, func() {
+			routeMetaData, err := s.App.ProtoRevKeeper.BuildTwoPoolRoute(
+				s.Ctx,
+				tc.swapDenom,
+				tc.tokenIn,
+				tc.tokenOut,
+				tc.poolId,
+			)
+
+			if tc.hasRoute {
+				s.Require().NoError(err)
+				s.Require().Equal(len(tc.expectedRoute), len(routeMetaData.Route.PoolIds()))
+
+				for index, trade := range tc.expectedRoute {
+					s.Require().Equal(trade.PoolId, routeMetaData.Route[index].PoolId)
+					s.Require().Equal(trade.OutputDenom, routeMetaData.Route[index].TokenOutDenom)
+				}
+			} else {
+				s.Require().Equal(len(tc.expectedRoute), len(routeMetaData.Route.PoolIds()))
 				s.Require().Error(err)
 			}
 		})